Move buttons chromatic test to the button components folder
