+++ packages/react-components/components/buttons/stories/imports.mdx
@@ -1,0 +1,3 @@
+```javascript
+    import { Button } from "@orbit-ui/react-buttons";
+```
