+++ packages/react-components/components/shared/src/utils/index.js
@@ -8,3 +8,4 @@ export * from "./defer";
 export * from "./keyboards";
 export * from "./handlers";
 export * from "./styles";
+export * from "./semantic-ui";
