+++ packages/semantic-ui-theme/semantic/src/definitions/elements/button.less
@@ -1109,9 +1109,10 @@
 .ui.fluid.button {
   width: 100%;
 }
-.ui.fluid.button {
+
+/*.ui.fluid.button {
   display: block;
-}
+}*/
 
 .ui.two.buttons {
   width: 100%;
