+++ packages/semantic-ui-theme/semantic/src/themes/sharegate/elements/button.overrides
@@ -6,6 +6,7 @@
     min-height: 40px;
     display: flex;
     align-items: center;
+    justify-content: center;
 }
 
 .ui.button.icon.tiny.circular {
