--- storybook/stories/semantic-ui/react/button/button.stories.mdx
@@ -1,280 +1,0 @@
-import { AddIcon, CalendarIcon24, SignoutIcon24 } from "@orbit-ui/icons";
-import { Button, Label } from "semantic-ui-react";
-import { Meta, Preview, Story } from "@storybook/addon-docs/blocks";
-import { SemanticLayout } from "../semantic-layout";
-import { createComponentSection, paramsBuilder } from "@utils";
-import Imports from "./imports.mdx";
-import Installation from "../installation.mdx";
-
-<Meta
-    title={createComponentSection("Button")}
-    parameters={paramsBuilder()
-        .docsLayout(<SemanticLayout documentationPath="/elements/button" />)
-        .build()}
-/>
-
-# Button
-
-## Installation
-
-<Installation imports={<Imports />} />
-
-## Usage
-
-### Default
-
-A default button.
-
-<Preview>
-    <Story name="default">
-        <Button>Cutoff</Button>
-    </Story>
-</Preview>
-
-### Active
-
-A button can be active.
-
-<Preview>
-    <Story name="active">
-        <Button active>Cutoff</Button>
-    </Story>
-</Preview>
-
-### Fluid
-
-A button can take the width of its container.
-
-<Preview>
-    <Story name="fluid">
-        <Button fluid>To infinity and beyond</Button>
-    </Story>
-</Preview>
-
-### Sizes
-
-A button can vary in size.
-
-<Preview>
-    <Story name="height short">
-        <div className="flex flex-row items-start">
-            <Button className="short">Short</Button>
-            <Button className="tall">Tall button</Button>
-        </div>
-    </Story>
-</Preview>
-
-### Disabled
-
-A button can be disabled.
-
-<Preview>
-    <Story name="disabled">
-        <Button disabled>Cutoff</Button>
-    </Story>
-</Preview>
-
-### Basic
-
-A basic button have a subtle appearance.
-
-<Preview>
-    <Story name="basic">
-        <div>
-            <div className="flex">
-                <Button basic>Cutoff</Button>
-                <Button basic primary>Cutoff</Button>
-            </div>
-            <div className="flex">
-                <Button basic disabled>Cutoff</Button>
-                <Button basic active>Cutoff</Button>
-                <Button basic primary disabled>Cutoff</Button>
-                <Button basic primary active>Cutoff</Button>
-            </div>
-        </div>
-    </Story>
-</Preview>
-
-### Emphasis
-
-A button can show different levels of emphasis.
-
-<Preview>
-    <Story name="emphasis">
-        <div>
-            <div className="flex">
-                <Button primary>Primary</Button>
-                <Button secondary>Secondary</Button>
-            </div>
-            <div className="flex">
-                <Button primary disabled>Primary</Button>
-                <Button secondary disabled>Secondary</Button>
-                <Button primary active>Primary</Button>
-                <Button secondary active>Secondary</Button>
-            </div>
-        </div>
-    </Story>
-</Preview>
-
-### Inverted
-
-A button can appear on a dark background.
-
-<Preview>
-    <Story name="inverted primary">
-        <div className="bg-marine-700 pa3">
-            <div className="flex">
-                <Button primary inverted>Cutoff</Button>
-            </div>
-            <div className="flex">
-                <Button primary inverted disabled>Cutoff</Button>
-                <Button primary inverted active>Cutoff</Button>
-            </div>
-        </div>
-    </Story>
-</Preview>
-
-### Ghost
-
-A ghost button is similar to a standard button, but have no basic colour until hovered.
-
-<Preview>
-    <Story name="ghost">
-        <div>
-            <div className="flex">
-                <Button className="ghost">Cutoff</Button>
-                <Button className="ghost" primary>Cutoff</Button>
-            </div>
-            <div className="flex">
-                <Button className="ghost" disabled>Cutoff</Button>
-                <Button className="ghost" active>Cutoff</Button>
-                <Button className="ghost" disabled primary>Cutoff</Button>
-                <Button className="ghost" active primary>Cutoff</Button>
-            </div>
-        </div>
-    </Story>
-</Preview>
-
-### Icon
-
-A button can contain an icon.
-
-<Preview>
-    <Story name="icon standard">
-        <div>
-            <div>
-                <Button icon><SignoutIcon24 className="w6 h6 mr1" />Click me</Button>
-            </div>
-            <div className="flex">
-                <Button icon disabled><SignoutIcon24 className="w6 h6 mr1" />Click me</Button>
-                <Button icon active><SignoutIcon24 className="w6 h6 mr1" />Click me</Button>
-            </div>
-        </div>
-    </Story>
-</Preview>
-
-### Icon Only
-
-A button can only contain an icon.
-
-<Preview>
-    <Story name="icon only">
-        <Button><AddIcon className="flex w6 h6"/></Button>
-    </Story>
-</Preview>
-
-### Circular
-
-A button can be circular.
-
-<Preview>
-    <Story name="circular">
-        <div className="flex flex-row items-start">
-            <Button icon circular size="tiny"><AddIcon className="flex w6 h6"/></Button>
-            <Button icon circular size="small"><AddIcon className="flex w6 h6"/></Button>
-            <Button icon circular>D</Button>
-            <Button icon circular size="large"><AddIcon className="flex w7 h7"/></Button>
-        </div>
-    </Story>
-</Preview>
-
-### Loading
-
-A button can show a loading indicator.
-
-<Preview>
-    <Story name="loading">
-        <div>
-            <div className="flex">
-                <Button loading>Loading</Button>
-                <Button loading primary>Loading</Button>
-                <Button loading basic>Loading</Button>
-                <Button loading basic primary>Loading</Button>
-            </div>
-            <div className="flex">
-                <Button disabled loading>Loading</Button>
-                <Button disabled loading primary>Loading</Button>
-                <Button disabled loading basic>Loading</Button>
-                <Button disabled loading basic primary>Loading</Button>
-            </div>
-            <div className="flex">
-                <Button active loading>Loading</Button>
-                <Button active loading primary>Loading</Button>
-                <Button active loading basic>Loading</Button>
-                <Button active loading basic primary>Loading</Button>
-            </div>
-        </div>
-    </Story>
-</Preview>
-
-### Success
-
-A button can hint towards a positive consequence.
-
-<Preview>
-    <Story name="success">
-        <div>
-            <div>
-                <Button positive>Cutoff</Button>
-            </div>
-            <div className="flex">
-                <Button positive disabled>Cutoff</Button>
-                <Button positive active>Cutoff</Button>
-            </div>
-        </div>
-    </Story>
-</Preview>
-
-### Negative
-
-A button can hint towards a negative consequence.
-
-<Preview>
-    <Story name="negative">
-        <div>
-            <div>
-                <Button negative>Cutoff</Button>
-            </div>
-            <div className="flex">
-                <Button negative disabled>Cutoff</Button>
-                <Button negative active>Cutoff</Button>
-            </div>
-        </div>
-    </Story>
-</Preview>
-
-### Groups
-
-Buttons can exist together as a group.
-
-<Preview>
-    <Story name="groups">
-        <div className="flex">
-            <Button.Group basic toggle>
-                <Button disabled type="button"><Label circular size="mini" empty color="red"/>One</Button>
-                <Button icon type="button"><CalendarIcon24 className="w6 h6 mr1"/>Two</Button>
-                <Button icon type="button" active><CalendarIcon24 className="w6 h6 mr1"/>Three</Button>
-            </Button.Group>
-        </div>
-    </Story>
-</Preview>
