--- storybook/stories/semantic-ui/react/button/imports.mdx
@@ -1,3 +1,0 @@
-```javascript
-    import { Button } from "semantic-ui-react";
-```
